tools/requirements.txt
zarr
fsspec[gcs]
transformers
datasets
hydra-core
